+++ NEWS.md
@@ -1,3 +1,12 @@
+# outils 0.0.0.9001
+
+## Major changes
+
+- result for `get_function_names` has been changed to a data.frame with
+
+  1. function name
+  2. path file name inside which function has been created
+
 # outils 0.0.0.9000
 
 ## Major changes